Disable mw.ui styles and rename main.less
Login form was weird and its styles needed to be completely reimplemented
as a result.

main.less is now screen.less so adding a mobile.less and whatnot later will
be less awkward.

Also removed the js for the menus since touch devices already have built-in
support for converting hover events to click events.

Change-Id: I6294c424a57f8dbb5efa9f89cb28eb19d40a2c49